@@ -3,6 +3,7 @@ description: A new Flutter project.
 
 publish_to: 'none'
 
+# Flutter Version :- 3.3.8
 
 version: 1.0.0+1
 
@@ -16,39 +17,42 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.3
-  shared_preferences: ^2.0.12
-  fluttertoast: ^8.0.8
-  permission_handler: ^8.3.0
-  syncfusion_flutter_gauges: 19.4.48
-  google_maps_flutter: ^2.1.1
-  location: ^4.3.0
-  geolocator: ^8.0.5
+  shared_preferences: ^2.0.15
+  fluttertoast: ^8.1.1
+  permission_handler: ^10.2.0
+  syncfusion_flutter_gauges: ^20.3.56
+  google_maps_flutter: ^2.2.1
+  location: ^4.4.0
+  geolocator: ^9.0.2
   flutter_polyline_points: ^1.0.0
   pedometer: ^3.0.0
-  lottie: ^1.2.1
-  flutter_rating_bar: ^4.0.0
-  floor: ^1.2.0
-  path_provider: ^2.0.8
+  lottie: ^2.0.0
+  flutter_rating_bar: ^4.0.1
+  floor: ^1.3.0
+  path_provider: ^2.0.11
   solid_bottom_sheet: ^0.1.10
   date_picker_timeline: ^1.2.3
   fl_chart: ^0.41.0
-  stop_watch_timer: ^1.3.1
-  date_format: ^2.0.5
-  flutter_local_notifications: ^9.2.0
+  stop_watch_timer: ^2.0.0
+  date_format: ^2.0.7
+  flutter_local_notifications: ^12.0.3+1
   flutter_native_timezone: ^2.0.0
   charts_flutter: ^0.12.0
-  share_plus: ^3.0.4
-  url_launcher: ^6.0.18
+  share_plus: ^6.3.0
+  url_launcher: ^6.1.6
   auto_size_text: ^3.0.0
-  rate_my_app: ^1.1.1+1
-  google_mobile_ads: ^1.0.1
-  app_tracking_transparency: ^2.0.2+1
+  rate_my_app: ^1.1.3
+  google_mobile_ads: ^2.3.0
+  app_tracking_transparency: ^2.0.3
+
+dependency_overrides:
+  platform: ^3.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  floor_generator: ^1.2.0
-  build_runner: ^2.1.5
+  floor_generator: ^1.3.0
+  build_runner: ^2.3.2
 
 
 flutter:
