@@ -511,7 +511,7 @@ class _DrinkWaterReminderScreenState extends State<DrinkWaterReminderScreen>
           const NotificationDetails(
             android: AndroidNotificationDetails('drink_water_reminder',
                 'Drink Water', channelDescription: 'This is reminder for drinking water on time',icon: 'ic_notification'),
-            iOS: IOSNotificationDetails(),
+            iOS: DarwinNotificationDetails(),
           ),
           androidAllowWhileIdle: true,
           uiLocalNotificationDateInterpretation:
